refactor: move the Option module to support/
`lib/components/` should hold only auto-importable Vue components, so the shared
Option types belong in `lib/support/` (alongside Day, File, etc.) rather than
`lib/components/Option.ts`. Repoint every importer.
