+++ lib/blocks/lens/components/LensFormFilterGroup.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import IconPlusCircle from '~icons/ph/plus-circle'
 import IconTrash from '~icons/ph/trash'
-import { type Option } from '../../../components/Option'
 import SButton from '../../../components/SButton.vue'
 import SInputSegments from '../../../components/SInputSegments.vue'
 import { useTrans } from '../../../composables/Lang'
+import { type Option } from '../../../support/Option'
 import { type FieldData } from '../FieldData'
 import LensFormFilterCondition, { type FilterCondition } from './LensFormFilterCondition.vue'
